+++ src/main/java/com/lec/spring/service/ProviderService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface ProviderService {
     List<ProvLodging> getLodgings(Long userId);
-    ProvLodging getAllDetails(int lodgingId);
+    ProvLodging getAllDetails(Long lodgingId);
     void saveLodging(ProvLodging lodging);
 
     List<ProvLodging> getLodgingsAndRoomsByUserId(Long userId);
